+++ src/components/InteractiveList/InteractiveList.types.part.tsx
@@ -104,6 +104,15 @@ export interface InteractiveListProps extends StandardProps {
    * Event transported from the wrapper in case of a click somewhere inside.
    */
   onClick?(e: React.MouseEvent<HTMLElement>): void;
+  /**
+   * Event fired when clicked outside of the component
+   */
+  onClickOutside?(): void;
+  /**
+   * Whether to position the list to top or bottom.
+   * Either `0` (bottom) or `1` (top). Make sure that the value provided is number.
+   */
+  direction?: InteractiveListDirection;
 }
 
 export interface InteractiveListState {