+++ webpack.config.js
@@ -68,7 +68,36 @@ module.exports = {
   module: {
     rules: [
       {
+        test: /node_modules\/((ansi-styles|strip-ansi|ansi-regex|react-dev-utils|chalk|typescript-plugin-inner-jsx)\/).*/,
+        use: {
+          loader: 'babel-loader',
+          options: {
+            cacheDirectory: true,
+            presets: [
+              [
+                '@babel/preset-env',
+                {
+                  targets: {
+                    ie: 11,
+                    esmodules: true,
+                  },
+                  modules: 'commonjs',
+                },
+              ],
+              '@babel/preset-typescript',
+              '@babel/preset-react',
+            ],
+            plugins: [
+              ['@babel/plugin-proposal-decorators', { legacy: true }],
+              ['@babel/plugin-proposal-class-properties', { loose: true }],
+              '@babel/plugin-transform-runtime',
+            ],
+          },
+        },
+      },
+      {
         test: /\.tsx$/,
+        exclude: /node_modules/,
         loader: 'awesome-typescript-loader',
         options: {
           getCustomTransformers: () => ({ before: [innerJsxTransformer] }),
@@ -76,10 +105,12 @@ module.exports = {
       },
       {
         test: /\.ts$/,
+        exclude: /node_modules/,
         loader: 'awesome-typescript-loader',
       },
       {
         enforce: 'pre',
+        exclude: /node_modules/,
         test: /\.js$/,
         loader: 'source-map-loader',
       },