+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # Precise UI Changelog
 
+## 0.6.0
+
+- Implement tag builder with auto-completion possibility
+- `Flyout` fixed compatibility with IE11
+- `Headline` styles update
+- Updated label in `Modal`
+- Added new icons
+- Remove disabled attribute from HTML element of `InteractiveSurface`
+- Bugfix: Close `DropdownField` if other one was opened (#44)
+- Bugfix: Fixed crash of `Rating` component (#69)
+- Added `onChangeDone` event to `Slider` component
+- Added `validationRules` for `Form`
+- Fixed: missed selection state of `RadioButtonGroup` in non-controlled mode with form context.
+- Compatibility with IE11
+- `Sidebar` component added
+- Fixed: component wrapped with `withValidation` hoc didn't display an error
+- Feature: new Dropdown props (open / direction)
+
 ## 0.5.0
 
 - Initial release